Restructure sync modules
