+++ core/src/utils/mod.rs
@@ -1,4 +1,3 @@
-pub mod racy_lock;
 pub mod sync;
 
 use alloc::vec::Vec;
@@ -6,8 +5,6 @@ use core::{
     fmt::Debug,
     ops::{Bound, Range},
 };
-#[cfg(feature = "std")]
-pub use std::sync::LazyLock;
 
 // RE-EXPORTS
 // ================================================================================================
@@ -15,8 +12,6 @@ pub use miden_crypto::utils::{
     collections, uninit_vector, ByteReader, ByteWriter, Deserializable, DeserializationError,
     Serializable, SliceReader,
 };
-#[cfg(not(feature = "std"))]
-pub use racy_lock::RacyLock as LazyLock;
 pub use winter_utils::group_slice_elements;
 #[cfg(feature = "std")]
 pub use winter_utils::ReadAdapter;
